+++ termsequel.cpp
@@ -110,6 +110,7 @@ void show_help(
         "\n OTHERS_PERMISSIONS Others permissions"
 #endif
         "\n LAST_MODIFICATION  The last modification of the file"
+        "\n CREATION_DATE      The creation date of the file"
         "\n RELATIVE_PATH      The relative path of the file"
         "\n ABSOLUTE_PATH      The absolute path of the file"
         "\nSQL instructions available"