+++ Changelog.md
@@ -76,3 +76,11 @@ Contains information about the versions of Termsequel
         | Add support for variables | [ISSUE]() | [PR](https://github.com/sgtcortez/Termsequel/pull/69) | Now, it is possible to use environment variables, and terminal variables(if supported by your terminal) with termsequel **FROM** source.  
         | Add CREATION_DATE column | [ISSUE](https://github.com/sgtcortez/Termsequel/issues/31) | [PR](https://github.com/sgtcortez/Termsequel/pull/77) | Now, the file creation date is available |
         | Add NOT EQUAL comparasion | [ISSUE](https://github.com/sgtcortez/Termsequel/issues/74) | [PR](https://github.com/sgtcortez/Termsequel/pull/78) | Now, it is possible to filter by not equal values |
+
+- Release 1.1.1
+
+    - Bug Fixes
+
+        | Bug | Issue | Pull Request | Description |
+        | :-:   | :-: | :-: | :-: |    
+